# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SPV API
  version: 1.0.0
  description: API para votos, puntos y analisis de planificacion.
servers:
  - url: http://localhost:8787
paths:
...
  /api/points/balance:
    get:
      summary: Balance del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Balance disponible e historico
  /api/points/balance/{userId}:
    get:
      summary: Balance por userId
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Balance del usuario
  /api/points/credit:
    post:
      summary: Acreditar puntos
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Credito aplicado
  /api/points/debit:
    post:
      summary: Debitar puntos
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Debito aplicado
        '409':
          description: Saldo insuficiente
  /api/points/transfer:
    post:
      summary: Transferir puntos entre usuarios
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Transferencia aplicada
        '409':
          description: Saldo insuficiente
  /api/points/convert:
    post:
      summary: Convertir puntos a item externo
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Conversion iniciada
...
  /api/webhooks/points/credit:
    post:
      summary: Webhook de acreditacion de puntos externos
      parameters:
        - in: header
          name: x-webhook-secret
          required: false
          schema:
            type: string
        - in: header
          name: x-webhook-timestamp
          required: false
          schema:
            type: string
          description: Timestamp unix (segundos) requerido cuando se usa HMAC.
        - in: header
          name: x-webhook-signature
          required: false
          schema:
            type: string
          description: Firma `sha256=<hex>` o `<hex>` para validar payload.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, amount]
              properties:
                userId:
                  type: string
                amount:
                  type: number
                sourceSystem:
                  type: string
                requestId:
                  type: string
                eventId:
                  type: string
                metadata:
                  type: object
      responses:
        '201':
          description: Credito externo aplicado
        '401':
          description: Secret o firma invalida
        '409':
          description: Webhook replay detectado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT